+++ lib/easytest/expectation.rb
@@ -49,12 +49,12 @@ module Easytest
       Matcher::ContainExactly.new(actual: actual, expected: expected_items, negate: negate).match!
     end
 
-    def to_raise(expected)
+    def to_raise(expected, with_message = nil)
       raise FatalError, "`to_raise` requires a block like `expect { ... }.to_raise`" unless block
       raise FatalError, "`not.to_raise` can cause a false positive, so use `to_raise_nothing` instead" if negate?
       raise FatalError, "`to_raise` requires a Class, String, or Regexp" unless [Class, String, Regexp].any? { expected.is_a? _1 }
 
-      Matcher::Raise.new(actual: block, expected: expected, negate: negate).match!
+      Matcher::Raise.new(actual: block, expected: expected, negate: negate, with_message: with_message).match!
     end
 
     def to_raise_nothing